Move regions enum to regions folder
